fix email notifications
